% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{EulerBending-package}
\alias{EulerBending}
\alias{EulerBending-package}
\title{EulerBending}
\description{
\tabular{ll}{ Package: \tab EulerBending\cr
              Type: \tab Package\cr
              Version: \tab 0.0.1\cr
              Date: \tab 2013-11-29\cr
              License: \tab GPL-2\cr
              LazyLoad: \tab yes\cr
              LazyData: \tab yes\cr }
}
\author{
Kevin M. Middleton (\url{middletonk@missouri.edu})
}
\keyword{package}

