\docType{data}
\name{glosso}
\alias{glosso}
\title{Glossophaga Radius}
\format{A \code{data.frame} consisting of:
\tabular{ll}{ Trial: \tab Trial number\cr
              P: \tab Load (N)\cr
              L: \tab Bone length\cr
              x_pix: \tab x position in pixels\cr
              u_pix: \tab x position in pixels}}
\source{
  Recorded by Kevin Middleton.
}
\description{
  Glossophaga radius in cantilever bending. Scale: 107
  pixels = 1 cm.
}
\keyword{datasets}

